Change all fetching data source to return Result

+++ data/src/main/java/app/tivi/data/repositories/popularshows/PopularShowsRepository.kt
@@ -17,6 +17,7 @@
 package app.tivi.data.repositories.popularshows
 
 import android.arch.paging.DataSource
+import app.tivi.data.entities.Success
 import app.tivi.data.repositories.shows.LocalShowStore
 import app.tivi.data.repositories.shows.ShowRepository
 import app.tivi.data.resultentities.PopularEntryWithShow
@@ -48,14 +49,15 @@ class PopularShowsRepository @Inject constructor(
     }
 
     private suspend fun updatePopularShows(page: Int, resetOnSave: Boolean) {
-        traktDataSource.getPopularShows(page, 20)
-                .map { (show, entry) ->
+        val response = traktDataSource.getPopularShows(page, 20)
+        when (response) {
+            is Success -> {
+                response.data.map { (show, entry) ->
                     // Grab the show id if it exists, or save the show and use it's generated ID
                     val showId = showStore.getIdOrSavePlaceholder(show)
                     // Make a copy of the entry with the id
-                    entry.copy(showId = showId)
-                }
-                .also { entries ->
+                    entry.copy(showId = showId, page = page)
+                }.also { entries ->
                     if (resetOnSave) {
                         localStore.deleteAll()
                     }
@@ -68,5 +70,7 @@ class PopularShowsRepository @Inject constructor(
                         }
                     }
                 }
+            }
+        }
     }
 }
